refactor: TurnEntity 관련 로직을 BoardDao에서 분리
- TurnDao에서 해당 로직을 관리한다.
